correcting bugs
