+++ src/br/unicamp/meca/system2/codelets/ExpectationCodelet.java
@@ -1,8 +1,15 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+/*******************************************************************************
+ * Copyright (c) 2018  DCA-FEEC-UNICAMP                  *
+ * All rights reserved. This program and the accompanying materials            *
+ * are made available under the terms of the GNU Lesser Public License v3      *
+ * which accompanies this distribution, and is available at                    *
+ * http://www.gnu.org/licenses/lgpl.html                                       *
+ *                                                                             *
+ * Contributors:                                                               *
+ *     W. Gibaut and R. R. Gudwin                                              * 
+ *                                                                             *
+ *                                                                             *
+ ******************************************************************************/
 package br.unicamp.meca.system2.codelets;
 
 import br.unicamp.cst.core.entities.Codelet;
@@ -16,7 +23,7 @@ import br.unicamp.cst.core.entities.Memory;
 
 /**
  *
- * @author wander
+ * @author W. Gibaut
  */
 public abstract class ExpectationCodelet extends Codelet{
     //transformar em abstrato
