+++ src/br/unicamp/meca/system2/codelets/LearningCodelet.java
@@ -1,8 +1,15 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+/*******************************************************************************
+ * Copyright (c) 2018  DCA-FEEC-UNICAMP                  *
+ * All rights reserved. This program and the accompanying materials            *
+ * are made available under the terms of the GNU Lesser Public License v3      *
+ * which accompanies this distribution, and is available at                    *
+ * http://www.gnu.org/licenses/lgpl.html                                       *
+ *                                                                             *
+ * Contributors:                                                               *
+ *     W. Gibaut and R. R. Gudwin                                              * 
+ *                                                                             *
+ *                                                                             *
+ ******************************************************************************/
 package br.unicamp.meca.system2.codelets;
 
 import br.unicamp.cst.core.entities.Codelet;
@@ -11,19 +18,13 @@ import br.unicamp.cst.representation.owrl.AbstractObject;
 import br.unicamp.cst.representation.owrl.Property;
 import br.unicamp.cst.representation.owrl.QualityDimension;
 //import com.google.common.primitives.Doubles;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import br.unicamp.meca.util.Encoder;
 
 /**
  *
- * @author wander
+ * @author W. Gibaut
  */
 public abstract class LearningCodelet extends Codelet{
